Encoder x4, Faster loop, add tim7 for debug

+++ Core/Inc/decoder.h
@@ -13,10 +13,10 @@
 
 // 定义轮子速度结构体
 typedef struct {
-    int wheel_FL; // 左前轮速度
-    int wheel_FR; // 右前轮速度
-    int wheel_RL; // 左后轮速度
-    int wheel_RR; // 右后轮速度
+    float wheel_FL; // 左前轮速度
+    float wheel_FR; // 右前轮速度
+    float wheel_RL; // 左后轮速度
+    float wheel_RR; // 右后轮速度
 } WheelSpeeds;
 
 // 计算麦克纳姆轮速度的函数声明
